+++ CfxGenerator/ApiTypes/CefStructType.cs
@@ -71,7 +71,7 @@ public class CefStructType : CefType {
     }
 
     public override string ProxySymbol {
-        get { return "ulong"; }
+        get { return "IntPtr"; }
     }
 
     public override string RemoteSymbol {
